--- PleaseAllow/PleaseAllow.h
@@ -1,19 +1,0 @@
-//
-//  PleaseAllow.h
-//  PleaseAllow
-//
-//  Created by Gagandeep Singh on 22/3/18.
-//  Copyright © 2018 Gagandeep Singh. All rights reserved.
-//
-
-#import <UIKit/UIKit.h>
-
-//! Project version number for PleaseAllow.
-FOUNDATION_EXPORT double PleaseAllowVersionNumber;
-
-//! Project version string for PleaseAllow.
-FOUNDATION_EXPORT const unsigned char PleaseAllowVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <PleaseAllow/PublicHeader.h>
-
-
