+++ Sources/PleaseAllow/PermissionManagers/PermissionManagerType.swift
