+++ Sources/PleaseAllow/PermissionManagers/PermissionStatus.swift
