+++ Sources/PleaseAllow/PleaseAllowResult.swift
