+++ Sources/PleaseAllow/SoftAsk/SoftAskView+Public.swift
@@ -6,7 +6,7 @@
 //  Copyright © 2018 Gagandeep Singh. All rights reserved.
 //
 
-import Foundation
+import UIKit
 
 extension SoftAskView {
