+++ Tests/LinuxMain.swift
@@ -1,0 +1,7 @@
+import XCTest
+
+import PleaseAllowTests
+
+var tests = [XCTestCaseEntry]()
+tests += PleaseAllowTests.allTests()
+XCTMain(tests)
