+++ Tests/PleaseAllowTests/XCTestManifests.swift
@@ -1,0 +1,9 @@
+import XCTest
+
+#if !canImport(ObjectiveC)
+public func allTests() -> [XCTestCaseEntry] {
+    return [
+        testCase(PleaseAllowTests.allTests),
+    ]
+}
+#endif
